Agentic AI Projects/Project 3/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
# psycopg2-binary==2.9.9  # Optional - only needed for PostgreSQL
python-dotenv==1.0.0
llama-index==0.9.20
llama-index-llms-openai==0.1.5
llama-index-embeddings-openai==0.1.4
openai==1.3.5
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.20
python-multipart==0.0.6
alembic==1.12.1
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0


